+++ Html/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 ===================
 
+* 4.0.2
+  * Fix: Render values that look empty for select options
+
+* 3.2.7
+  * Fix: Render values that look empty for select options
+
 * 3.2.2
   * Fix: Use HTMLElement as default context when rendering